move deposits from member to wallet
